ciao-down: Report cloud-init errors
Any errors reported while executing the cloud-init runcmds are now reported
back to ciao-down and the user.  Errors do not currently interrupt the
preparation of the VM.  They merely inform the user that some aspect of the
VM setup may have failed and that Single VM might not run correctly.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>