ciao-down: Intelligently choose resources for VM
This PR modifies ciao-down to intelligently choose resources for the VM
it creates and boots.  These defaults are computed from the total amount
of memory available on the host and the number of cores it has.  These
values are divided by 2 to give the values that will be assigned to the
guest.

The user can also override these values where preparing and starting a
VM.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>