ciao-down: Ensure VM gets deleted when prepare fails
There were a couple of race conditions preventing the instance directory
getting deleted when prepare failed, either through an error or because
it had been cancelled.  We were not waiting for the VM to actually exit
and also we were quitting the process before the routine that deleted the
instance directory actually got around to doing so.  This commit fixes
both issues.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>