ciao-down: Use apt section of cloudinit file
The cloudinit file has a dedicated section for configuring apt but
we weren't using it.  Instead we were executing all of the apt
commands in the runcmd and write_files section of the cloud-init file.
This commit moves the instructions that set up the apt proxies and
initiate the apt-get upgrade to the apt section.  The other commands
such as setting up sources and installing packages can not be used
as these sections do not handle proxies correctly.

Fixes #1131

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>