ciao-down: Only use ciao-down ssh key
This commit modifies the ciao-down connect command to only use the
identity key generated by ciao-down.  We've had some reports that
users with a lot of keys in their .ssh directories are unable to
connect to a ciao-down VM, as ssh seems to be trying these keys
in preference to the key supplied by ciao-down, and is giving
up after a number of connection attempts have failed.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>