ciao-down: Add --package-upgrade prepare flag
This commit adds a new flag to the prepare command, --package-upgrade,
that can be used to provide a hint to workloads indicating whether
packages contained within the base image should be updated or not
during the first boot.  Updating packages can be quite time consuming
and may not be necessary if the user just wants to create a throw away
VM to test something under a particular distro.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>