ciao-down: Ensure skeleton files are copied
There was a bug in some of the ciao-down workloads that prevented the
skeleton files, such as .bashrc, from being copied into the user accounts
created on the guest.  This happened if a mount was specified with a path
in the user's home directory.  The issue has been fixed by delaying the
mounting of files until the user's home directory has been created.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>