ciao-down: save workload in instance directory
Previously, ciao-down only saved the dynamic instance data in the
instance directory.  The reason is that this is the only information
that is actually needed by ciao-down after the instance has been
created.  However, it might be useful for users to have a record
of the exact workload used to create their instances as there's
no guarantee that the original workload template will continue to
exist after the instance has been created.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>