ciao-down: Rename instanceSpec to workloadSpec
instanceSpec is misleading as the content of this structure is actually
what doesn't change per-instance. Rename it.

To be consistent, this commit also renames the insSpec field and
variables to just spec.

In the future it's likely we'll fold the per-instance "instance" struct
(and the insData field of workload) into workloadSpec.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>