ciao-down: Fold instance data into the main workloadSpec
At the moment all of our workload files have an empty 2nd document to be
used to save per-instance data. This is a bit awkward for people that
want to write workload files. Removing it will remove the need to
explain the little quirk.

First step is to consolidate the per-VM data (instance data) into the
workload struct in order to, in a following commit, load the full
workloadSpec from one document and remove the middle document from the
YAML file.

Signed-off-by: Damien Lespiau <damien.lespiau@intel.com>