Fix a small bug with the proxy handling
The special address, 10.0.2.2, used by the guest to communicate with the
host, was not added to the no_proxy environment variable.  This meant that
communication with the host via http could fail.  An example of this failure
is when trying to use the download function as a non-root user inside the
cloud-init script.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>