CCloudvm no longer uses template.Must
When ccloudvm was first written it supported only one workload, ciao.
This workload was built into the ccloudvm binary as a Go string.  When
parsing the workload, template.Must was used.  This made sense as any
syntax errors in the template were indicative of a problem with
ccloudvm itself.  Now however, workloads are user supplied and so can
often contain errors.  Panicing on error is no longer the correct
thing to do.  It also means that the VM under construction does not
get cleaned up correctly.

Fixes: https://github.com/intel/ccloudvm/issues/14

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>