Allow user to specify the size of the rootfs
Up until now ccloudvm assigns an arbitrary maximum size of 60 GiBs to
the rootfs of the newly created VM.  The user was unable to override
this value, either in the workload file or on the command line.  This
commit allows the user to do both of these things.  The default is still
60GiB if the user does nothing.

Fixes: https://github.com/intel/ccloudvm/issues/9

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>