ccvm: Implement workload inheritance
Implement and test workload inheritance allowing a workload to inherit
from another workload.

One change that comes from this is that workload defaults are now not
injected as part of unmarshal but instead in a default specification
that is always merged in.

The SSH port mapping however is special cased (as before) to ensure that
it is always mapped to allow ccloudvm connectivity.

For the cloud-config data, templates are applied and then the
cloud-config data is merged; with sequences in the yaml document being
appended (from parent towards child) and maps being merged at the top
level.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>