Use a new muxer for each new creation request
ccloudvm creates a local web server when it creates a new instance.
The instance communicates with this web server providing information
about the progress of the instance creation and also to request
ccloudvm to download images on its behalf.  This web server used
to use the default mux.  Going forward this is not going to work.
When ccvm is running as a service it is going to support multiple
creation requests during its life time.  If each of these creation
requests registers a handler on the default mux ccvm will panic.  This
commit then introduces a new mux per creation request.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>