Switch ccloudvm over to a client/server model
This commit transforms the ccvm package into a systemd service.  The
ccloudvm setup command installs this service.  All other ccloudvm
commands issued from the command line send a request to the ccvm
service over a domain socket.  The service is socket activated and
exits if there are no clients using its services.  The service is
run as a user service.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>