ccvm: Remove nil values from parsed cloud-config
If a cloud-config is parsed with an empty block (as a result of
templating) the yaml unmarshaller will create a key with a value of nil
as it does not know if the type should be a scalar or a collection. When
marshalling this will then result in the string "nil" being output as
the scalar (string) value for the field which is invalid.

Instead strip out nil values from the top-level config after parsing.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>