Remove the downloadCh global variable
The downloadCh variable that kept track of the download channel so that
other parts of ccvm could initiate downloads, has been removed.  The
downloadCh is still created when ccvm starts up but is now passed to the
parts of the code that needs it via a parameter.  This increases the
number of arguments that we need to pass to certain functions but it
avoids the need to have a global variable.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>