Use reflect.Select in the server main loop
This change is in preparation for when the server go routine needs to
manage multiple instance go routines.  Using reflect.Select will allow
us to listen on an arbritrary number of channels at the same time in
the same go routine.  We can thus safely detect the death of an instance
go routine, either because it has been deleted or it failed during
creation.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>