Add the instanceLoop
This commit adds a new go routine that manages the lifecyle of an instance
and serialises all commands sent to that instance.  The server go routine
manages the instance go routines, detects when they quit, because the
instance has been deleted, or failed to start, and instructs them to
quit when the ccvm service shuts down.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>