Change default memory and cpu allocations
CCloudvm used to assign half of the host's available memory and cpus to a
new instance, if the workload or the command line does not specify otherwise.
These figures may have been appropriate in the past when ccloudvm was used
to create a single instance of a ciao development environment.  Now that
it supports multiple instances, these defaults no longer make sense.  The
new defaults are 1GB of RAM and 1 vcpu.  This commit changes the ciao and
clear-container workloads to ensure they have enough RAM to work.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>