Remove hostname from the workload specification
Hostnames can no longer be specified in the workload specification, which makes
sense really, as you're unlikely to want multiple instances with the same
hostname.  Instead the hostname of the instance is derived from the name of the
instance.  As there are restrictions on the characters that can appear in hostnames
the instance naming criteria has been tightened up.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>