Fix bug in download manager
If you started ccloudvm and then deleted an image that was stored in the
cache using rm, ccloudvm would not be able to use that image until the
service was restarted.  ccloudvm would think the file was downloaded when
in actual fact it was not.  Any VM based on this file would fail to
start as the backing image did not in fact exist.  This commit fixes the
issue by checking the file really exists every time a download request
is received.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>