Use random port for the local http server
The host and guest communicate with each other over a http server
bound to a port on 127.0.0.1.  Previously, this port was hardcoded
which probably wasn't great from a security point of view but it
also meant that two instances could not be created at the same time
as each instance tried to create its own webserver and bind it to
the same port.  The ports for local web servers are now randomly
assigned by the OS.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>