Fix data race and enable race detection in semaphore
This commit fixes a data race in the code that manages VM installation
and enables race detection in the automated semaphore builds.  The code
is a lot more concurrent that it used to be with the introduction of the
download manager and the instanceLoop so we really need to enable race
detection now.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>