Read GOPATH from user's current shell
The GOPATH variable was read by the ccvm service rather than from
the shell in which the ccloudvm create command was executed.  This
mean that the user was unable to specify a different GOPATH.  Also
the GOPATH didn't seem to be correctly set inside the service anyway.
This commit fixes the issue by passing the user's GOPATH to the
ccvm service along with all the other data in the create command.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>