Fix a race condition in the download manager
There was a race condition in the download manager in which it signaled
that a download had been completed after the file had been downloaded
but before it had been moved to its final location.  This was causing
some of the unit tests to fail, which were being told that a download
had completed but when they went to check it, they discovered that the
file did not exist.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>