Fix an issue with ccloudvm start
ccloudvm start was broken.  It gave an error about host-ip not being
defined.  The issue was that the MergeCustom function was not reading
the host-ip assigned during the create command from the state.yaml
file.  This was done to avoid reading the host ip from a workload
defintion.  However, such a value would not be used in practice as ccvm will
automatically assign an IP address at creation time and this IP address
will override anything specified in the workload definition.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>