Fix some issues with the sempahore.sh file
There were some typos and some indentation problems.  There was also
a bug in which the script made assumptions about the IP address on
which guest ports were mapped.  These assumptions are no longer valid
since the introduction of multiple instances.  The script now executes
the ccloudvm status command to determine the host ip address on which
guest ports are exposed.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>