Semaphore.sh now runs when there are existing instances
Previously, sempahore.sh failed if the user had existing ccloudvm instances.
This wasn't a problem when running in sempahore but it made it a bit awkward
to run the script locally as one needed to destroy all their existing instances,
before running semaphore.sh.  This commit fixes the issue by referring to the
instance it creates by name.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>