Derive instance host IP address from uid
The host IP address assigned to an instance if the user does not
specifically request an ip address at creation time is now derived
from the user's uid.  This avoids port clashes which are almost
guaranteed to happen if more than one user uses cccloudvm on the
same machine at the same time.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>