Fix service shutdown issues
The cloudvm service shutdown could be blocked by an uncompleted transaction.
There was also another issue in which an RPC request entering the server
could block if a signal to shutdown the service occured at the same time.
If the service go routine received the signal first and shut down, writing
to the actionCh would block forever.  The issues are fixed by ensuring that
the resultCh is closed when a transaction has been processed and draining
any data from the result channels of each active transaction when shutting
down the service.  We also uses select in api.go when sending commands down
the actionCh, so that we can quit if there's a signal.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>