Change permissions of the domain socket
The domain socket used by the ccloudvm command line tool and the systemd
unit service was created with permissions of 0666, which was not restrictive
enough.  We don't want other users to communicate with our ccloudvm service.
This commit restricts permissions on the domain socket so that it can only
be accessed by the current user.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>