Support EFI Boot
This commit allows the user to specify that an image should be booted
with a particular BIOS.  To boot with a given BIOS the user needs to
provide a URL to the BIOS of choice in the workload.  The http,
https and file schemes are supported.  The BIOS is copied into the
instance directory.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>