Improve example conistency in README.md
The README.md file provides a number of examples of how to use the
ccloudvm command.  However, it was very inconistent in its
presentation of the command.  Sometimes ccloudvm was written as
./ccloudvm, sometimes as ccloudvm or sometimes as $ ccloudvm.
This commit makes the use of the command consistent and standardises
on $ ccloudvm, which is the easiest of the three to read.

Fixes: https://github.com/intel/ccloudvm/issues/58

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>