Add debug connection information
With the addition of multiple instance support it became difficult to
debug buggy cloud-init files.  The reason is that if cloud-init fails
to complete, the creation command will also not complete.  As ccloudvm
serialises the commands it sends to a given instance, no other commands,
e.g., connect, can be issued on that instance until create has completed,
making it difficult to connect to instances with buggy cloud-init files
to debug them.

This commit works around the issue by providing the user with ssh
connection details for the instance being created if they pass
the --debug option to create.  Connect's still not going to work but
at least the user will have a way to connect to their VM to debug it.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>