remove erroneous download notifications
Everytime you create a ccloudvm instance at least two messages appear
on the screen informing the user that the backing image is being downloaded.
But most of the time this isn't true.  If the image is already stored
in ccloudvm's cache it is not downloaded again.  This commit modifies the
download code to ensure that messages pertaining to downloading of files
are only presented to the user when something is being donwloaded.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>