Ensure 127.0.0.1 is part of the no_proxy env variable
This commit ensures that 127.0.0.1 is present in the
workspace.NoProxy field if proxies are enabled on the host.  The
chances are that if the host has proxies enabled, 127.0.0.1 will be
present in its no_proxy anyway, but we might as well make sure.  If
it is present, we'll take care not to add it twice.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>