Add IP address of eth0 port to no_proxy list
Current no_proxy list is populated with hostname and few other IPs. All qemu
vms are currently coming up with IP 10.0.2.15 whcih is not listed in that list.
This causes issues when a tool works off ip addresses. for eg: kubeadm
installation will hang and timeout eventually since it is not able to reach
10.0.2.15 which gets routed out of the system. This is an awful hack until we
can identify the IP on the NIC and inject it during VM provisioning time.

Signed-off-by: Ganesh Maharaj Mahalingam <ganesh.mahalingam@intel.com>