Remove CRI-O and K8S from clearcontainers-3.x.yaml
This commit removes the commands that install CRI-O and K8S from the
clearcontainers-3.x workload.  This is done as the CRI-O installation
doesn't work and also as K8S and CRI-O don't really belong in this
workload.  Should there be a need to have these components in the
future a new workload that inherits from clearcontainers-3.x will need
to be created.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>