Fix workload search
ccloudvm was unable to find the workload directory if the directory
resided in a GOPATH that was not the default.  The reason was that
the search is performed using the environment variables of the service
and not that of the client.  Although the client's GOPATH is passed
into the service it was not being used when searching for workloads.
This has now been fixed:

Fixes: https://github.com/intel/ccloudvm/issues/87

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>