Disable service timeout when not using systemd
The ccvm service starts a timer when it has completed executing its last
command and its command queue is empty.  The timer lasts for one minute
and if no new commands are received during this minute, the service
shuts down.  This makes sense when using systemd as the service will be
automically restarted to service any new commands by systemd socket
activation.  However, when running ccvm without systemd there's nothing
to restart it, so it doesn't make sense to shut it down when its idle.

Signed-off-by: Mark Ryan <mark.d.ryan@intel.com>